+++ src/QueryType/Update/RequestBuilder.php
@@ -128,10 +128,10 @@ class RequestBuilder extends BaseRequestBuilder
     {
         $xml = '<delete>';
         foreach ($command->getIds() as $id) {
-            $xml .= '<id>'.htmlspecialchars($id, ENT_NOQUOTES).'</id>';
+            $xml .= '<id>'.$this->getHelper()->escapeXMLCharacterData($id).'</id>';
         }
         foreach ($command->getQueries() as $query) {
-            $xml .= '<query>'.htmlspecialchars($query, ENT_NOQUOTES).'</query>';
+            $xml .= '<query>'.$this->getHelper()->escapeXMLCharacterData($query).'</query>';
         }
         $xml .= '</delete>';
 
@@ -234,7 +234,7 @@ class RequestBuilder extends BaseRequestBuilder
         } elseif ($value instanceof \DateTimeInterface) {
             $value = $this->getHelper()->formatDate($value);
         } else {
-            $value = htmlspecialchars($value, ENT_NOQUOTES);
+            $value = $this->getHelper()->escapeXMLCharacterData($value);
         }
 
         $xml .= '>'.$value.'</field>';